pyproject.toml
[project]
name = "chatbot-api"
version = "0.1.0"
description = ""
authors = [{ name = "Gabriel Lima", email = "grtl@ecomp.poli.br" }]
requires-python = ">=3.13, <4.0"
readme = "README.md"

dependencies = [
    "fastapi[standard]==0.115.9",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-auth (>=2.39.0,<3.0.0)",
    "google-api-python-client (>=2.167.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "anthropic (>=0.50.0,<0.51.0)",
    "ollama (>=0.4.8,<0.5.0)",
    "langchain>=0.2.17",
    "langchain-community>=0.2.19",
    "pypdf>=5.6.0",
    "chromadb==1.0.9",
    "sentence-transformers>=4.1.0",
    "langchain-huggingface>=0.0.3",
    "langchain-text-splitters>=0.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5,<9",
    "pytest-cov>=6.1.1,<7",
    "taskipy>=1.14.1,<2",
    "ruff>=0.11.6,<0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.3.5,<9",
    "pytest-cov>=6.1.1,<7",
    "taskipy>=1.14.1,<2",
    "ruff>=0.11.6,<0.12",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'